More cleanup, added missing interface docs
