+++ sql/tables.go
@@ -14,60 +14,70 @@
 
 package sql
 
-// Table represents the backend of a SQL table.
+import "fmt"
+
+// Table is a SQL table.
 type Table interface {
 	Nameable
-	String() string
+	fmt.Stringer
+	// Schema returns the table's schema.
 	Schema() Schema
+	// Collation returns the table's collation.
 	Collation() CollationID
+	// Partitions returns the table's partitions in an iterator.
 	Partitions(*Context) (PartitionIter, error)
+	// PartitionRows returns the rows in the given partition, which was returned by Partitions.
 	PartitionRows(*Context, Partition) (RowIter, error)
 }
 
-// Table2 is an experimental future interface alternative to Table to provide faster access.
-type Table2 interface {
-	Table
-
-	PartitionRows2(ctx *Context, part Partition) (RowIter2, error)
-}
-
-// TableFunction is a node that is generated by a function
+// TableFunction is a node that is generated by a function and can be used as a table factor in many SQL queries.
 type TableFunction interface {
 	Node
 	Expressioner
 	Databaser
 	Nameable
 
-	// NewInstance returns a new instance of the table function
-	NewInstance(ctx *Context, db Database, expressions []Expression) (Node, error)
+	// NewInstance calls the table function with the arguments provided, producing a Node
+	NewInstance(ctx *Context, db Database, args []Expression) (Node, error)
 }
 
-// TemporaryTable allows tables to declare that they are temporary
+// TemporaryTable allows tables to declare that they are temporary (created by CREATE TEMPORARY TABLE).
+// Only used for validation of certain DDL operations -- in almost all respects TemporaryTables are indistinguishable 
+// from persisted tables to the engine.
 type TemporaryTable interface {
+	// IsTemporary should return true if the table is temporary to the session
 	IsTemporary() bool
 }
 
-// TableWrapper is a node that wraps the real table. This is needed because
-// wrappers cannot implement some methods the table may implement.
+// TableWrapper is a node that wraps the real table. This is needed because wrappers cannot implement some methods the
+// table may implement. This interface is used in analysis and planning and is not expected to be implemented by
+// integrators.
 type TableWrapper interface {
 	// Underlying returns the underlying table.
 	Underlying() Table
 }
 
-// FilteredTable is a table that can produce a specific RowIter
-// that's more optimized given the filters.
+// FilteredTable is a table that can filter its result rows from RowIter using filter expressions that would otherwise
+// be applied by a separate Filter node. 
 type FilteredTable interface {
 	Table
+	// Filters returns the filter expressions that have been applied to this table.
 	Filters() []Expression
+	// HandledFilters returns the subset of the filter expressions given that this table can apply.
 	HandledFilters(filters []Expression) []Expression
+	// WithFilters returns a table with the given filter expressions applied.
 	WithFilters(ctx *Context, filters []Expression) Table
 }
 
-// ProjectedTable is a table that can produce a specific RowIter
-// that's more optimized given the columns that are projected.
+// ProjectedTable is a table that can return only a subset of its columns from RowIter. This provides a very large 
+// efficiency gain during table scans. Tables that implement this interface must return the projected column in future 
+// calls to Schema
 type ProjectedTable interface {
 	Table
+	// WithProjections returns a version of this table with only the subset of columns named. Calls to Schema must
+	// only include these columns.
 	WithProjections(colNames []string) Table
+	// Projections returns the names of the column projections applied to this table.
 	Projections() []string
 }
 
@@ -106,14 +116,14 @@ type IndexAlterableTable interface {
 	RenameIndex(ctx *Context, fromIndexName string, toIndexName string) error
 }
 
-// ForeignKeyTable is a table that can declare its foreign key constraints, as well as be referenced.
+// ForeignKeyTable is a table that declares foreign key constraints, and can be referenced by other tables' foreign 
+// key constraints.
 type ForeignKeyTable interface {
 	IndexAddressableTable
 	// CreateIndexForForeignKey creates an index for this table, using the provided parameters. Indexes created through
 	// this function are specifically ones generated for use with a foreign key. Returns an error if the index name
 	// already exists, or an index on the same columns already exists.
 	CreateIndexForForeignKey(ctx *Context, indexDef IndexDef) error
-
 	// GetDeclaredForeignKeys returns the foreign key constraints that are declared by this table.
 	GetDeclaredForeignKeys(ctx *Context) ([]ForeignKeyConstraint, error)
 	// GetReferencedForeignKeys returns the foreign key constraints that are referenced by this table.
@@ -130,7 +140,13 @@ type ForeignKeyTable interface {
 	GetForeignKeyEditor(ctx *Context) ForeignKeyEditor
 }
 
-// CheckTable is a table that can declare its check constraints.
+// ForeignKeyEditor is a TableEditor that is addressable via IndexLookup.
+type ForeignKeyEditor interface {
+	TableEditor
+	IndexAddressable
+}
+
+// CheckTable is a table that declares check constraints.
 type CheckTable interface {
 	Table
 	// GetChecks returns the check constraints on this table.
@@ -147,6 +163,12 @@ type CheckAlterableTable interface {
 	DropCheck(ctx *Context, chName string) error
 }
 
+// PrimaryKeyTable is a table with a primary key.
+type PrimaryKeyTable interface {
+	// PrimaryKeySchema returns this table's PrimaryKeySchema
+	PrimaryKeySchema() PrimaryKeySchema
+}
+
 // PrimaryKeyAlterableTable represents a table that supports primary key changes.
 type PrimaryKeyAlterableTable interface {
 	Table
@@ -157,18 +179,7 @@ type PrimaryKeyAlterableTable interface {
 	DropPrimaryKey(ctx *Context) error
 }
 
-type PrimaryKeyTable interface {
-	// PrimaryKeySchema returns this table's PrimaryKeySchema
-	PrimaryKeySchema() PrimaryKeySchema
-}
-
-// TableEditor is the base interface for sub interfaces that can update rows in a
-// table during an INSERT, REPLACE, UPDATE, or DELETE statement.
-type TableEditor interface {
-	RowReplacer
-	RowUpdater
-}
-
+// EditOpenerCloser is the base interface for table editors, and deals with statement boundaries.
 type EditOpenerCloser interface {
 	// StatementBegin is called before the first operation of a statement. Integrators should mark the state of the data
 	// in some way that it may be returned to in the case of an error.
@@ -179,11 +190,6 @@ type EditOpenerCloser interface {
 	// StatementComplete is called after the last operation of the statement, indicating that it has successfully completed.
 	// The mark set in StatementBegin may be removed, and a new one should be created on the next StatementBegin.
 	StatementComplete(ctx *Context) error
-}
-
-type AutoIncrementEditor interface {
-	AutoIncrementSetter
-	AutoIncrementGetter
 }
 
 // InsertableTable is a table that can process insertion of new rows.
@@ -205,7 +211,7 @@ type RowInserter interface {
 	Closer
 }
 
-// DeleteableTable is a table that can process the deletion of rows
+// DeletableTable is a table that can delete rows.
 type DeletableTable interface {
 	Table
 	// Deleter returns a RowDeleter for this table. The RowDeleter will get one call to Delete for each row to be deleted,
@@ -220,11 +226,12 @@ type RowDeleter interface {
 	// each row to process for the delete operation, which may involve many rows. After all rows have been processed,
 	// Close is called.
 	Delete(*Context, Row) error
-	// Close finalizes the delete operation, persisting the result.
+	// Closer finalizes the delete operation, persisting the result.
 	Closer
 }
 
-// TruncateableTable is a table that can process the deletion of all rows.
+// TruncateableTable is a table that can process the deletion of all rows either via a TRUNCATE TABLE statement or a
+// DELETE statement without a WHERE clause. This is usually much faster that deleting rows one at a time. 
 type TruncateableTable interface {
 	Table
 	// Truncate removes all rows from the table. If the table also implements DeletableTable and it is determined that
@@ -233,10 +240,8 @@ type TruncateableTable interface {
 	Truncate(*Context) (int, error)
 }
 
-// AutoIncrementTable is a table that supports AUTO_INCREMENT.
-// Getter and Setter methods access the table's AUTO_INCREMENT
-// sequence. These methods should only be used for tables with
-// and AUTO_INCREMENT column in their schema.
+// AutoIncrementTable is a table that supports AUTO_INCREMENT. Getter and Setter methods access the table's
+// AUTO_INCREMENT sequence. These methods should only be used for tables with and AUTO_INCREMENT column in their schema.
 type AutoIncrementTable interface {
 	Table
 	// GetNextAutoIncrementValue gets the next AUTO_INCREMENT value. In the case that a table with an autoincrement
@@ -253,21 +258,20 @@ type AutoIncrementTable interface {
 type AutoIncrementSetter interface {
 	// SetAutoIncrementValue sets a new AUTO_INCREMENT value.
 	SetAutoIncrementValue(*Context, uint64) error
-	// Close finalizes the set operation, persisting the result.
+	// Closer finalizes the set operation, persisting the result.
 	Closer
 }
 
+// AutoIncrementGetter is implemented by tables that support AUTO_INCREMENT to return the next value that will be
+// inserted, given a particular insert value provided by the client.
 type AutoIncrementGetter interface {
 	GetNextAutoIncrementValue(ctx *Context, insertVal interface{}) (uint64, error)
-	// Close finalizes the set operation, persisting the result.
-	Closer
 }
 
-// RowReplacer is a combination of RowDeleter and RowInserter.
-type RowReplacer interface {
-	EditOpenerCloser
-	RowInserter
-	RowDeleter
+// AutoIncrementEditor is an interface for tables that support changing the value of auto increment columns.
+type AutoIncrementEditor interface {
+	AutoIncrementSetter
+	AutoIncrementGetter
 }
 
 // ReplaceableTable allows rows to be replaced through a Delete (if applicable) then Insert.
@@ -276,6 +280,13 @@ type ReplaceableTable interface {
 	// Replacer returns a RowReplacer for this table. The RowReplacer will have Insert and optionally Delete called once
 	// for each row, followed by a call to Close() when all rows have been processed.
 	Replacer(ctx *Context) RowReplacer
+}
+
+// RowReplacer is a combination of RowDeleter and RowInserter.
+type RowReplacer interface {
+	EditOpenerCloser
+	RowInserter
+	RowDeleter
 }
 
 // UpdatableTable is a table that can process updates of existing rows via update statements.
@@ -293,6 +304,13 @@ type RowUpdater interface {
 	Update(ctx *Context, old Row, new Row) error
 	// Closer finalizes the update operation, persisting the result.
 	Closer
+}
+
+// TableEditor is the combination of interfaces that allow any table edit operation:
+// i.e. INSERT, UPDATE, DELETE, REPLACE
+type TableEditor interface {
+	RowReplacer
+	RowUpdater
 }
 
 // RewritableTable is an extension to Table that makes it simpler for integrators to adapt to schema changes that must
@@ -315,18 +333,6 @@ type RewritableTable interface {
 	RewriteInserter(ctx *Context, oldSchema, newSchema PrimaryKeySchema, oldColumn, newColumn *Column, idxCols []IndexColumn) (RowInserter, error)
 }
 
-// UnresolvedTable is a Table that is either unresolved or deferred for until an asOf resolution
-type UnresolvedTable interface {
-	Nameable
-	// Database returns the database name
-	Database() string
-	// WithAsOf returns a copy of this versioned table with its AsOf
-	// field set to the given value. Analogous to WithChildren.
-	WithAsOf(asOf Expression) (Node, error)
-	// AsOf returns this table's asof expression.
-	AsOf() Expression
-}
-
 // AlterableTable should be implemented by tables that can receive ALTER TABLE statements to modify their schemas.
 type AlterableTable interface {
 	Table
@@ -341,9 +347,22 @@ type AlterableTable interface {
 	ModifyColumn(ctx *Context, columnName string, column *Column, order *ColumnOrder) error
 }
 
-// ForeignKeyEditor is a TableEditor that is addressable via IndexLookup.
-type ForeignKeyEditor interface {
-	TableEditor
-	IndexAddressable
+// UnresolvedTable is a Table that is either unresolved or deferred for until an asOf resolution. 
+// Used by the analyzer during planning, and is not expected to be implemented by integrators. 
+type UnresolvedTable interface {
+	Nameable
+	// Database returns the database name
+	Database() string
+	// WithAsOf returns a copy of this versioned table with its AsOf
+	// field set to the given value. Analogous to WithChildren.
+	WithAsOf(asOf Expression) (Node, error)
+	// AsOf returns this table's asof expression.
+	AsOf() Expression
 }
 
+// Table2 is an experimental future interface alternative to Table to provide faster access.
+type Table2 interface {
+	Table
+
+	PartitionRows2(ctx *Context, part Partition) (RowIter2, error)
+}
